PET/requirements.txt
pandas>=1.5
numpy>=1.23
scikit-learn>=1.2
scipy>=1.9
Pillow>=9.0
matplotlib>=3.6
